feat: unification of HighLevel
