+++ README.md
@@ -18,4 +18,4 @@ We split the problem into an
     - they take at least the arguments ax and **kwargs. ax is assumed to be either None or a matplotlib Axes. If ax is None then the matplotlib function `gca` is invoked to retrieve the current active Axes.
 - Each class definition is seperated by two lines
 - Respect the 80 column limit in all code files. Exceptions:
-    - property comments 
+    - property comments, which start at the 80 column mark
