+++ unittests/Tests.py
@@ -1,10 +1,14 @@
-from Experiment import Experiment
-from HighLevel import GlobalPathPlanner
-from World import * 
+
+# external imports
 import matplotlib.pyplot as plt
 import numpy as np  
 from scipy.spatial import Voronoi, voronoi_plot_2d
 from typing import Tuple
+
+# internal imports
+from Experiment import Experiment
+from GlobalPlanning import GlobalPathPlanner
+from World import * 
 from Agent import *
 from experiments.experiment_small import plot_results
 
@@ -17,17 +21,6 @@ def shift_time():
     traj = Trajectory(np.array([[-10,20,0.2,0.01,-1.0]]), np.array([0,1,2,3,4]))
     traj.shiftTime(1)
     assert(np.all(traj.t == np.array([1,2,3,4,5])))                 
-
-# Nice seeds for 10 sets and 0.5 fraction: 235
-def generate_experiment(n_sets=15,fraction=0.5,seed=784) -> Experiment:
-    if seed is not None:
-        np.random.seed(seed)
-    ex = Experiment()
-    ex.AddRandomVoronoiPoints(n_sets)
-    ex.GeneratePartitioning()
-    ex.AddRandomTargets(fraction=fraction)
-    ex.AssignRandomAgent()
-    return ex
 
 def test_random_boundary_point(n_sets = 10):
     ex = Experiment()
@@ -52,50 +45,50 @@ def test_voronoi(M = 10):
     return
 
 def test_plot_world(n_sets=10):
-    ex = generate_experiment(n_sets=n_sets)
+    ex = Experiment.generate(n_sets=n_sets)
     fig, ax = ex.PlotWorld()
     return ex  
 
 def test_dist_to_boundary(n_sets=10):
-    ex = generate_experiment(n_sets=n_sets)
+    ex = Experiment.generate(n_sets=n_sets)
     fig, ax = ex.PlotWorld()
     ex._world.plotDistToBoundary(ax)
     return ex
 
 def test_travel_cost():
-    ex = generate_experiment()
+    ex = Experiment.generate()
     fig, ax = ex.PlotWorld()
     ex._world.plotTravelCostPerRegion(ax)
     return ex
     
-def test_rrt(max_iter = 1000, n_sets=20, plot = True) -> GlobalPathPlanner:
-    ex = generate_experiment(n_sets=n_sets)
+def test_rrt(max_iter = 100, n_sets=20, plot = True) -> GlobalPathPlanner:
+    ex = Experiment.generate(n_sets=n_sets)
     assert(isinstance(ex, Experiment))
     gpp = GlobalPathPlanner(ex._world)
-    gpp._plot_options.toggle_all_plotting(plot)
+    gpp._plot_options.toggleAllPlotting(plot)
     gpp._plot_options._par = False
     gpp._plot_options._psr = False
     fig, ax = ex.PlotWorld()
     gpp.planPath(ex._world.target(1).p(), ex._world.target(9).p(), max_iter, ax)
-    ex._world.plotTravelCostPerRegion(ex, ax)
+    ex._world.plotTravelCostPerRegion(ax)
     plt.tight_layout()
     return gpp
 
 def test_tsp(n_sets=20, plot = False) -> GlobalPathPlanner:
-    ex = generate_experiment(n_sets=n_sets)
+    ex = Experiment.generate(n_sets=n_sets, fraction=0.2)
     assert(isinstance(ex, Experiment))
     gpp = GlobalPathPlanner(ex._world)
-    gpp._plot_options.toggle_all_plotting(plot)
-    fig, ax = ex.PlotWorld(ax)
-    gpp.solveTSP(ax)
+    gpp._plot_options.toggleAllPlotting(plot)
+    fig, ax = ex.PlotWorld()
+    gpp.solveTSP()
     # gpp.tsp.plotTargetDistances(ax)
     gpp.plotTSPSolution(ax, color='red', linewidth=2)
-    po = ex._agent.plotSensorQuality(ex, ax)
+    po = ex._agent.plotSensorQuality(ax)
     gpp.plotTSPSolution(ax, color='red', linewidth=2)
     return gpp
 
 def test_local_controller(n_sets=20) -> Experiment:
-    ex = generate_experiment(n_sets=n_sets)
+    ex = Experiment.generate(n_sets=n_sets)
     assert(isinstance(ex, Experiment))
     
     target = ex._world.targets()[0]
@@ -118,7 +111,7 @@ def test_local_controller(n_sets=20) -> Experiment:
     return ex
 
 def test_cycle(n_sets=20) -> Experiment:
-    ex = generate_experiment(n_sets=n_sets)
+    ex = Experiment.generate(n_sets=n_sets)
     assert(isinstance(ex, Experiment))
     
     target = ex._world.targets()[0]
@@ -131,7 +124,7 @@ def test_cycle(n_sets=20) -> Experiment:
     return ex
 
 def test_bilevel_optimization(n_sets=20) -> Experiment:
-    ex = generate_experiment(n_sets=n_sets)
+    ex = Experiment.generate(n_sets=n_sets)
     assert(isinstance(ex, Experiment))
     
     target = ex._world.targets()[0]
